transferred an array of films
