+++ project/src/pages/main-page/main-page.tsx
@@ -1,42 +1,8 @@
 import FilmListItem from '../../components/film-list-item/film-list-item';
 import { IPromo } from '../../index';
+import useGetFilms from '../../hooks/useGetfilms';
 
 function MainPage({promoFilm}:{promoFilm: IPromo}): JSX.Element {
-  const filmsList = [
-    {
-      title: 'Fantastic Beasts: The Crimes of Grindelwald',
-      src: 'img/fantastic-beasts-the-crimes-of-grindelwald.jpg',
-    },
-    { title: 'Bohemian Rhapsody', src: 'img/bohemian-rhapsody.jpg' },
-    { title: 'Macbeth', src: 'img/macbeth.jpg' },
-    { title: 'Aviator', src: 'img/aviator.jpg' },
-    {
-      title: 'We need to talk about Kevin',
-      src: 'img/we-need-to-talk-about-kevin.jpg',
-    },
-    {
-      title: 'What We Do in the Shadows',
-      src: 'img/what-we-do-in-the-shadows.jpg',
-    },
-    { title: 'Revenant', src: 'img/revenant.jpg' },
-    { title: 'Johnny English', src: 'img/johnny-english.jpg' },
-    { title: 'Shutter Island', src: 'img/shutter-island.jpg' },
-    { title: 'Pulp Fiction', src: 'img/pulp-fiction.jpg' },
-    {
-      title: 'No Country for Old Men',
-      src: 'img/no-country-for-old-men.jpg',
-    },
-    { title: 'Snatch', src: 'img/snatch.jpg' },
-    { title: 'Moonrise Kingdom', src: 'img/moonrise-kingdom.jpg' },
-    { title: 'Seven Years in Tibet', src: 'img/seven-years-in-tibet.jpg' },
-    { title: 'Midnight Special', src: 'img/midnight-special.jpg' },
-    { title: 'War of the Worlds', src: 'img/war-of-the-worlds.jpg' },
-    { title: 'Dardjeeling Limited', src: 'img/dardjeeling-limited.jpg' },
-    { title: 'Orlando', src: 'img/orlando.jpg' },
-    { title: 'Mindhunter', src: 'img/mindhunter.jpg' },
-    { title: 'Midnight Special', src: 'img/midnight-special.jpg' },
-  ];
-
   return (
     <>
       <section className='film-card'>
@@ -177,7 +143,7 @@ function MainPage({promoFilm}:{promoFilm: IPromo}): JSX.Element {
           </ul>
 
           <div className="catalog__films-list">
-            {filmsList.map((film) => (
+            {useGetFilms().map((film) => (
               <FilmListItem key={film.title} title={film.title} src={film.src} />
             ))}
           </div>
